throw customerCrudException problems occurred

+++ src/Customer/CustomerCrud.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Ecurring\WooEcurring\Customer;
 
+use WP_User;
+
 /**
  * Customer crud using WP user meta as a storage.
  */
@@ -19,6 +21,14 @@ class CustomerCrud implements CustomerCrudInterface
      */
     public function saveEcurringId(int $localUserId, string $ecurringId): void
     {
+        if (! $this->userExists($localUserId)) {
+            throw new CustomerCrudException(
+                sprintf(
+                    'Couldn\'t save eCurring customer id for user %1$d because this user doesn\'t exist',
+                    $localUserId
+                )
+            );
+        }
         update_user_meta($localUserId, self::ECURRING_CUSTOMER_ID_STORAGE_KEY, $ecurringId);
     }
 
@@ -27,6 +37,15 @@ class CustomerCrud implements CustomerCrudInterface
      */
     public function getEcurringId(int $localUserId): string
     {
+        if (! $this->userExists($localUserId)) {
+            throw new CustomerCrudException(
+                sprintf(
+                    'Couldn\'t get eCurring customer id for user %1$d because this user doesn\'t exist',
+                    $localUserId
+                )
+            );
+        }
+
         return (string) get_user_meta($localUserId, self::ECURRING_CUSTOMER_ID_STORAGE_KEY, true);
     }
 
@@ -35,6 +54,15 @@ class CustomerCrud implements CustomerCrudInterface
      */
     public function saveMollieMandateId(int $localUserId, string $mollieMandateId): void
     {
+        if (! $this->userExists($localUserId)) {
+            throw new CustomerCrudException(
+                sprintf(
+                    'Couldn\'t save Mollie mandate id for user %1$d because this user doesn\'t exist',
+                    $localUserId
+                )
+            );
+        }
+
         update_user_meta($localUserId, self::MOLLIE_MANDATE_ID_STORAGE_KEY, $mollieMandateId);
     }
 
@@ -43,6 +71,16 @@ class CustomerCrud implements CustomerCrudInterface
      */
     public function getMollieMandateId(int $localUserId): string
     {
+
+        if (! $this->userExists($localUserId)) {
+            throw new CustomerCrudException(
+                sprintf(
+                    'Couldn\'t get Mollie mandate id for user %1$d because this user doesn\'t exist',
+                    $localUserId
+                )
+            );
+        }
+
         return get_user_meta($localUserId, self::MOLLIE_MANDATE_ID_STORAGE_KEY, true);
     }
 
@@ -51,7 +89,36 @@ class CustomerCrud implements CustomerCrudInterface
      */
     public function clearCustomerData(int $localUserId): void
     {
-        delete_user_meta($localUserId, self::ECURRING_CUSTOMER_ID_STORAGE_KEY);
-        delete_user_meta($localUserId, self::MOLLIE_MANDATE_ID_STORAGE_KEY);
+        if (! $this->userExists($localUserId)) {
+            throw new CustomerCrudException(
+                sprintf(
+                    'Couldn\'t clear customer data for user %1$d because this user doesn\'t exist',
+                    $localUserId
+                )
+            );
+        }
+
+        $result = delete_user_meta($localUserId, self::ECURRING_CUSTOMER_ID_STORAGE_KEY) &&
+                    delete_user_meta($localUserId, self::MOLLIE_MANDATE_ID_STORAGE_KEY);
+
+        if (! $result) {
+            sprintf(
+                'Couldn\'t clear customer data. Please, check %1$s and %2$s meta ' .
+                'fields of user %3$d manually and delete them if needed.',
+                self::ECURRING_CUSTOMER_ID_STORAGE_KEY,
+                self::MOLLIE_MANDATE_ID_STORAGE_KEY,
+                $localUserId
+            );
+        }
+    }
+
+    /**
+     * Check if user with given ID exists.
+     *
+     * @return bool True for user exists, false otherwise.
+     */
+    protected function userExists(int $userId): bool
+    {
+        return WP_User::get_data_by('id', $userId) !== false;
     }
 }
